+++ apps/namada-interface/src/services/extensionEvents/handlers/namada.ts
@@ -2,12 +2,12 @@ import { Dispatch } from "react";
 
 import { chains } from "@namada/chains";
 import { Namada } from "@namada/integrations";
+import { TxType, TxTypeLabel } from "@namada/shared";
 
 import { addAccounts, fetchBalances } from "slices/accounts";
 import { actions as notificationsActions } from "slices/notifications";
 import { getToast, Toasts } from "slices/transfers";
 import { fetchValidators } from "slices/StakingAndGovernance/actions";
-import { TxType } from "@namada/shared";
 
 export const NamadaAccountChangedHandler =
   (dispatch: Dispatch<unknown>, integration: Namada) =>
@@ -36,14 +36,23 @@ export const NamadaUpdatedStakingHandler =
 export const NamadaTxStartedHandler =
   (dispatch: Dispatch<unknown>) => async (event: CustomEventInit) => {
     const { msgId, txType } = event.detail;
-    dispatch(notificationsActions.txStartedToast({ id: msgId, txType }));
+    dispatch(
+      notificationsActions.txStartedToast({
+        id: msgId,
+        txTypeLabel: TxTypeLabel[txType as TxType],
+      })
+    );
   };
 
 export const NamadaTxCompletedHandler =
   (dispatch: Dispatch<unknown>) => async (event: CustomEventInit) => {
     const { msgId, txType, success } = event.detail;
     dispatch(
-      notificationsActions.txCompletedToast({ id: msgId, txType, success })
+      notificationsActions.txCompletedToast({
+        id: msgId,
+        txTypeLabel: TxTypeLabel[txType as TxType],
+        success,
+      })
     );
   };
